[WIP] Scala Stream Collector: integrate the size violation bad row type (closes #4177)

+++ 2-collectors/scala-stream-collector/core/src/main/scala/com.snowplowanalytics.snowplow.collectors.scalastream/sinks/Sink.scala
@@ -25,7 +25,7 @@ import org.slf4j.LoggerFactory
 trait Sink {
 
   // Maximum number of bytes that a single record can contain
-  val MaxBytes: Long
+  val MaxBytes: Int
 
   lazy val log = LoggerFactory.getLogger(getClass())
